+++ src/main/java/kitchenpos/eatinorders/application/OrderService.java
@@ -2,10 +2,10 @@ package kitchenpos.eatinorders.application;
 
 import kitchenpos.deliveryorders.infra.KitchenridersClient;
 import kitchenpos.eatinorders.domain.*;
-import kitchenpos.eatinorders.infra.OrderRepository;
-import kitchenpos.eatinorders.infra.OrderTableRepository;
+import kitchenpos.eatinorders.domain.OrderRepository;
+import kitchenpos.eatinorders.domain.OrderTableRepository;
 import kitchenpos.menus.domain.Menu;
-import kitchenpos.menus.infra.MenuRepository;
+import kitchenpos.menus.domain.MenuRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
