+++ src/main/java/kitchenpos/eatinorders/domain/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.infra;
+package kitchenpos.eatinorders.domain;
 
 import kitchenpos.eatinorders.domain.OrderTable;
