+++ src/main/java/kitchenpos/menus/infra/JpaMenuGroupRepository.java
@@ -1,6 +1,7 @@
 package kitchenpos.menus.infra;
 
 import kitchenpos.menus.domain.MenuGroup;
+import kitchenpos.menus.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
