+++ src/main/java/kitchenpos/products/infra/JpaProductRepository.java
@@ -1,6 +1,7 @@
 package kitchenpos.products.infra;
 
 import kitchenpos.products.tobe.domain.Product;
+import kitchenpos.products.tobe.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
