+++ src/test/java/kitchenpos/eatinorders/application/InMemoryOrderRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.eatinorders.application;
 
 import kitchenpos.eatinorders.domain.Order;
-import kitchenpos.eatinorders.infra.OrderRepository;
+import kitchenpos.eatinorders.domain.OrderRepository;
 import kitchenpos.eatinorders.domain.OrderStatus;
 import kitchenpos.eatinorders.domain.OrderTable;
