+++ src/test/java/kitchenpos/eatinorders/application/InMemoryOrderTableRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.eatinorders.application;
 
 import kitchenpos.eatinorders.domain.OrderTable;
-import kitchenpos.eatinorders.infra.OrderTableRepository;
+import kitchenpos.eatinorders.domain.OrderTableRepository;
 
 import java.util.*;
