+++ src/test/java/kitchenpos/eatinorders/application/OrderServiceTest.java
@@ -1,10 +1,10 @@
 package kitchenpos.eatinorders.application;
 
 import kitchenpos.eatinorders.domain.*;
-import kitchenpos.eatinorders.infra.OrderRepository;
-import kitchenpos.eatinorders.infra.OrderTableRepository;
+import kitchenpos.eatinorders.domain.OrderRepository;
+import kitchenpos.eatinorders.domain.OrderTableRepository;
 import kitchenpos.menus.application.InMemoryMenuRepository;
-import kitchenpos.menus.infra.MenuRepository;
+import kitchenpos.menus.domain.MenuRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
