+++ src/test/java/kitchenpos/eatinorders/application/OrderTableServiceTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.eatinorders.application;
 
-import kitchenpos.eatinorders.infra.OrderRepository;
+import kitchenpos.eatinorders.domain.OrderRepository;
 import kitchenpos.eatinorders.domain.OrderStatus;
 import kitchenpos.eatinorders.domain.OrderTable;
-import kitchenpos.eatinorders.infra.OrderTableRepository;
+import kitchenpos.eatinorders.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
