+++ src/test/java/kitchenpos/products/application/InMemoryProductRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.products.application;
 
 import kitchenpos.products.tobe.domain.Product;
-import kitchenpos.products.infra.ProductRepository;
+import kitchenpos.products.tobe.domain.ProductRepository;
 
 import java.util.*;
 import java.util.stream.Collectors;
